Hack in specifying the host header
Way back we had an issue with a user's http requests not resolving
correctly due to, we believe, some stale caching of dns. To mitigate
that we performed the dns lookup ourselves and used the ip directly.

All fine and good, but our load balancer wants the host header to be
set to the host uri, not the ip. This library doesnt support that so
we are hacking it in in the simplest way possible.